backend/btc: track output index, simplify code
The tx proposal arg payment request had a pointer to the transaction
output it is attached to.

This is the wrong location to add this pointer, as the tx proposal
args don't contain this info - it is left empty when parsing the args
in btc/handlers.

The proper location for it is in the tx proposal itself. However, it's
simpler to just generically track the index of the recipient
output. Pointers seem brittle (a hard copy would break this). We also
only support one recipient output, and this is not likely to change.

Furthermore, we want to attach more metadata to the recipient output
later (silent payment address), so one way to track the index for all
cases is simpler.