frontend: allow action buttons to grow on mobile
With different translations the action buttons (send, receive, buy
& sell and WC) did not break nicely on small screen.

Due to limited button height the text could be outside of the button
and only visible in darkmode but not in normal mode (white on gray).

Changed to let the buttons break onto 2 lines of text if there is
not enough horozontal space.