frontend/transactions: show historical fiat value in txs list
In the account transaction list we used to show the fiat equivalent of
each tx amount, using the latest fiat rates. We decided that it was not
the best option, and could confuse users looking at past txs.

This changes the behavior showing the historical fiat amount instead. If
the historical amount is not available (e.g. tx not confirmed yet or
historical rates not updated), and the tx is recent enough, it will
make an estimation using the latest rates.

In case of estimation, the fiat conversion will have a `≈` char prefix.

If the historical fiat rates are not available and the tx is older, a
`---` will be shown, as in the tx details popup.