frontend/qt: revert d68b673ed, restore custom libserver.h
Using the auto-generated server/libsever.h directly does not work
after all, at least not when upgrading to Qt 6.2 on Windows with
MSVC.

cgo uses mingw gcc and produces C99 code in the header. The MSVC
compiler does not accept this any longer:

```
typedef _Fcomplex GoComplex64;
typedef _Dcomplex GoComplex128;
```

Something about syntax error, I guess _FComplex is not available when
compiling this as C++ with MSVC.

Not clear why it worked with Qt 5.15 - probably the compiler flags
have changed in Qt 6.2.

The custom libserver.h contains the same as the auto-generated one,
but reduced to the declarations we actually need, not all the Go C
preamble stuff that causes the compat problems with MSVC.