frontend/qt: update to Qt6 on macOS
brew's qt@6 would work excpt that it is missing all of the
qt@6/libexec folder, which contains the rcc tool we need to bundle the
frontend assets. Instead we use `aqtinstall` which installs Qt
binaries from official mirrors and includes all we need.

We get rid of osx-brew.sh which is only referenced in the BUILD.md and
instead inline the instructions, as it is strange to install brew and
other unprompted. Devs can follow along the BUILD.md instructions.

This commit also introduces universal builds for macOS. They are fat
binaries/libraries that include code for both the amd64 and arm64
architectures, so the app can run natively. As a result, the resulting
app size is roughly double to before.

Alternative considered but dismissed:
- Make and distribute two separate BitBox.app outputs per
  architecture. This reduces the app size, but is a bit more error
  prone for users and incurs more effort to distribute correctly. Also
  the AppStore requires universal builds, if we ever want to distbute
  the app through the AppStore.