frontend: fix back button with UseBackButton
The Android back button stopped working after the back button was used
to close a dialog. The reason is that the backbutton pop handler did
not remove the handler, as `slice(index, 1)` does not remove the
element at `index`, it in fact keeps it. When properly removing the
handler, back button behavior works as expected again.