frontend: keep fees dropdown always enabled
The fees dropdown show useful information such as expected
time, or with config.frontend.expertFee current cost sat/vB.

This change enables opening the dropdown to see the current fees.

Added disabled to the 'custom' option so it is only selectable if
there is valid txinput data (such as address and amount).