frontend: fix chart edgecase to properly update the total amount
On mobile the total changes when the user selects a point on the
chart so that they can see exact amount at the time.

With scrolling or 2nd finger changing the currency it can happen
that the amount does not update, ending up with a wrong amount/unit.

Fixed by resetting the tooltip value if the currency changes.