frontend: enable clipboard read/write for android/qt
This enables the use of the `navigator.clipboard` JS api to read/write
the clipboard on both Qt and Android and in particular:
- fixes the copy buttons in the Pocket order confirmation page
- updates the api of the CopyableInput component

The overriding of `navigator.clipboard.readText()` on Android is a bit
hacky, but without it the function raises a `Read permission denied`
error. The clipboard read feature is not currently used in the code,
but:
- with this we could implement some frontend improvement, e.g. a button
  to paste an address in the send view
- we need this in the LN branch to paste invoices (we used a workaround
  in the past to overcome the issue there) and splitting the commit
  seems quite overkill.

Please also note that reading from clipboard is disabled by default in
the iframes.