frontend: integrate btc direct buy widget
This is the first step of the BTC Direct integration and includes
the buy (fiat-to-coin) widget.

The widget requires linking to a CSS and loading a script from
BTC Direct as well as a few paramenters such as api-key.

To keep intergrations consistent and have a clean separation
between app and services, this commit includes a static page that
can be iframed, as BTC Direct does not provide an iframe solution.

Iframes have their own DOM and scoped CSS, so that our CSS cannot
accidentally mess up BTC Directs styles.

Currently data is only passed to the iframe via HTML attributes
but when needed we can use postMessage to communicate back to
the main app.

Disclaimers etc have just been copied from BTC Direct OTC info
and are currently placeholders, as the new disclaimers are not
yet final.