backend/btc: use PSBT as the internal format for passing unsigned tx
PSBT BIP: https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki

bitbox02-api-go now has a `BTCSignPSBT()` function, which simplifies
the signing code in the app significantly.

Benefits include:
- Standardized, so easier for devs to understand and maintain
- Future-proofing (interopable, extensible)
- Clear lifecycle:
  - Create (unsigned)
  - Update (add metadata)
  - Sign (BitBox)
  - Extract (create sigscripts and witnesses, create final signed tx)

The btcd/psbt lib is used and contains a function to extract the tx,
which makes our own functions to create sigscripts/witnesses
redundant.

godot is removed as a linter b/c of false positives and it not being
worth wasting time over not finishing a docstring with a dot.