tests: ensure that account is eventually synced.
With ensureAddresses running in a separate goroutine, it might happen
that after a call to account.Initialize() we perform checks that fail
because the ensureAddresses goroutine hasn't completed yet.

By adding a call to require.Eventually we ensure that the account gets
eventually synced (within a second) before proceeding with the rest
of the test.