ios: only detect the device in the backend once paired
The device first connects. Pairing happens later when a secured
characteristic is used. We assume the device is paired when we can
read the product string, at which point the backend register the
device. Otherwise, the app would launch into the device workflows in
the UI and get overlaid with the pairing code suddenly, which looks broken.