backend/handlers: fix header sync tx timestamp
The use of creation timestamp when the tx timestamp was not available
caused the app to show the current date for all the btc/ltc transactions
during the header syncing. This removes the use of the tx creation
timestamp, as it doesn't seem to be really useful anyway.

Now during header sync the tx date is not showed at all.