frontend: cleanup and remove unused sidebar related code
In the early versions of the BitBoxApp the sidebar had to be
activelly hidden. This was needed for example during the setup
to hide the sidebar.

https://github.com/BitBoxSwiss/bitbox-wallet-app/commit/737c7051af5e911cc267f48fec11da0f7571729c#diff-dfe1a46131ddaf4054e57c25f15a90037702158127d8959711cbcab68fac0418R161

In newer versions the setup view just covers everything so there
is no need to actively hide the sidebar.

Most related code has already been removed, but there was still a
case in BitBox01 setup and the code in the context, which can now
be removed.