ios: allow FaceID in screen lock
WIthout the NSFaceIDUsageDescription Info.plist key, only the passcode
could be used. With the key present, the app will ask permission to use
FaceID, and if granted, use that, and if not granted, use the passcode.

If you re-try the FaceID auth check twice without success, auth falls
back to passcode automatically.