backend: add argument to force highest fee rate in tx proposal
Previously, the highest fee rate was automatically applied to all
transaction proposals with a SLIP-24 payment request. For selling
and spending integrations that do not use SLIP-24 payment requests,
there should also be an option to trigger this same behaviour.

This commit replaces the previous conditioning on payment requests
with a general boolean field that is set independently and can be
used for other integrations. This also allows setting a custom fee
rate for transaction proposals with payment requests, if needed.

The TTxInput type excludes illegal combinations of this setting and
feeTarget or customFee, as either the highest fee should automatically
be used, or an explicitly specified fee target.