frontend: remove code for swiping the sidebar
In recent commit the hamburger navigation was replaced by bottom
menu navigation on mobile. Note: we still use the hamburger on
medium sized screens.

See a6d6af08191fc1b8c40c1352e62a6f5d1548df4c

The sidebar contained some code to listen for swipe events to open
and close the sidebar on mobile. So when a user swipes from left to
right in order to go back, it could happen that the swipe listener
takes over and opens the sidebar instead of going back 1 step in
history.

Removed the swipe listener in this commit.