frontend: add unlock software keystore in advanced settings
The hot software keystore is useful in Testnet on iOS and other
platforms.

This commit adds a new advanced settings item to unlock a testnet
wallet for testing purposes.

SettingsItem component can now show custom icon instead of the
default chevron to display an eject icon, to disconnect the test
wallet.