backend/btc/address: be explicit about chainIndex/addressIndex
The relative keypath always has two elements, the first element
indicating change and the second indicating the address index.

A relative keypath of arbitrary length is harder to handle. E.g. it is
not compatible with output descriptors with multipath
indices (`xpub/<0;1>/*`), where also two elements are required to
derive an address (the multipath index selecting between `<X;Y>` and the address index).

The unit tests change because the GetAddress test function used an
empty relative path instead of a chainIndex and addressIndex, which
now was forced to change with the stricter types.

I considered keeping `ChainIndex uint32` instead of `Change bool`,
because in multipaths, one could have more than two,
e.g. `xpub/<X;Y;Z>/*`, but I prefer the stricter type and it's not
likely we will ever need more than two (receive & change).