frontend: block ui in case of sign request
The bitbox02 only can handle one request at a time, this means that an
ongoing message signing request blocks the device until it is finished.

Currently, it is possible to request a message signing inside the pocket
workflow and leave the workflow while the signing request is ongoing.
If the user tried to make other operations, e.g. verifying an address
or opening the manage device screen, each bb02 request remains pending.

By blocking everything but the iframe the user is forced to stay in the
workflow and has to either complete or abort the signing.

When blocking the view should always surpress native back, by using

    <UseDisableBackButton />