frontend: fix sats unit in all-accounts tab
When sats mode is enabled, the all accounts tab on mobile showed the
value in satohis but with the wrong unit. It used the account unit
instead of the balance unit.

`hasAvailable` does not need to be checked, it means balance is `>0`,
but we show the balance regardless of it is 0 or larger.