frontend: scroll up after accepting (widget) terms
currently, after accepting a widget's terms the scroll
position will stay in the bottom. This is not ideal because
to use the widget we'd need to manually scroll up. This happens
because the route did not change when the screen changes from
the terms 'screen' to the widget 'screen'.

This commit intends to ensure that after every click on the
accept btn of any widget terms, the scrollable container scrolls
up, mimicking a scroll reset as if the user were to change route.