android: fix USB error handling to prevent crashes on disconnect
This fixes the Android app crashing or misbehaving upon disconnect if
it was in the middle of a read/write operation.

Validate bulkTransfer() result to throw an exception if it indicates
an error. Before, it would return a negative number to Go, which
would then interpret it as the number of bytes written (or in case of
read, an empty read buffer instead of an error).

An error happens here if the device is disconnected, but before the
backend learns about it (race condition).