frontend/ios: fix blank screens when foregrounding after inactivity
As part of system memory and energy management, iOS may terminate
the web view of the app at whim, e.g. after backgrounding the app
for a long time, usually combined with many different apps open during
active usage.

Whenever this happens, the "webViewWebContentProcessDidTerminate" hook
is triggered and forces a reload of the web view, set to the root page.
As this is only triggered if the web view was killed, it will not
interfere with regular usage.

Users may experience the reload briefly when they foreground the app
and the conditions for the reload are met. While this fixes the blank
screen issue, this is still somewhat suboptimal, as the user would
expect the last state from when they used the app to reload.