backend: fix off-by-one error in accounts scan
Hidden accounts are added and scanned so they can be shown if they are
not empty. This should have happened for the first 5 accounts
always (since we allowed 5 empty accounts to be added), but further
accounts only if the previous account was non-empty.

`maybeAddHiddenUnusedAccounts()` had an off-by-one error where we added and scanned
the sixth account even if all accounts before were empty. It should
have stopped at the fifth.

As a result, adding a new account worked even if the user already had
5 accounts, even if the last account was empty, as adding an account
implicitly assumes that existing hidden accounts only could exist if
they could be added.

We can fix the off-by-one, but since there might be users out
there (even if unlikely) that added such a sixth account, we now
should always auto-discover the first six instead of the first five
accounts, even if all are empty, so we bump this number. This should
not put more load on our servers, as the app already always scanned
six accounts due to this bug.