ios/android: fix screen lock bugs
On some Android versions (e.g. Android 9) the authentication flow was
causing an endless loop of auth requests. This was fixed by moving the
setup of `BiometricAuthHelper` from `onStart` to `onCreate`.

In both Android and iOS the auth flow freezed if there was no
authentication method setup on the device. This was fixed by refactoring
the authentication flow code and adding a new possible authentication
response `authres-missing`, that allows to handle this case.