backend/aopp: fix potential panic on uninitialized accounts
`GetUnusedReceiveAddresses()` returned `nil` on an unsynced account
now. Before, it used to block until it was synced. This change lead to
a regression in AOPP, where accessing `GetUnusedReceiveAddresses()[0]`
panicked. This would happen when one would enter the AOPP workflow and
unlocked the BitBox (which makes the account available) during the
AOPP workflow.

This commit adds an error return type to `GetUnusedReceiveAddresses()`
to make it clear this case should be handled. For AOPP, the quick fix
here is to wait while its synced like before.